requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
jinja2
python-multipart
# Добавьте драйвер для вашей БД. Для PostgreSQL:
psycopg2-binary
# Для SQLite (встроенный, но для некоторых версий Python может понадобиться)
# sqlite3